requirements.txt
attrs==21.4.0
bcrypt==3.2.0
boto3==1.21.20
botocore==1.24.20
cffi==1.15.0
click==8.0.4
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.10
flask-expects-json==1.7.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.32
urllib3==1.26.9
Werkzeug==2.0.3
